requirements.txt
alembic==1.7.7
Babel==2.9.1
boto3==1.21.27
botocore==1.24.28
certifi==2021.10.8
coverage==6.2
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Scss==0.5
Flask-SQLAlchemy==2.5.1
Flask-UUID==0.2
Flask-WTF==1.0.0
Flask-Session==0.4.0
Jinja2==3.0.3
redis==4.3.4
jsonpickle~=2.1.0
notifications-python-client==6.3.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyScss==1.3.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
urllib3==1.26.8
waitress==2.1.2
Werkzeug==2.1.1
WTForms==3.0.0
pandas==1.4.3
